ref(control_silo): Move UserOption model to users module
Part of moving control silo user related resources into the users module. Includes adding import shim for any getsentry refs.

Apart of (#73856)
