+++ src/sentry/api/endpoints/user_notification_email.py
@@ -7,8 +7,8 @@ from sentry.api.api_owners import ApiOwner
 from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import control_silo_endpoint
 from sentry.api.bases.user import UserEndpoint
-from sentry.models.options.user_option import UserOption
 from sentry.models.useremail import UserEmail
+from sentry.users.models.user_option import UserOption
 
 INVALID_EMAIL_MSG = (
     "Invalid email value(s) provided. Email values must be verified emails for the given user."
