+++ src/sentry/models/options/__init__.py
@@ -1,8 +1,9 @@
+from sentry.users.models.user_option import UserOption
+
 from .option import ControlOption, Option
 from .organization_option import OrganizationOption
 from .project_option import ProjectOption
 from .project_template_option import ProjectTemplateOption
-from .user_option import UserOption
 
 __all__ = (
     "Option",
