+++ src/sentry/users/models/user.py
@@ -339,10 +339,10 @@ class User(Model, AbstractBaseUser):
         from sentry.models.authidentity import AuthIdentity
         from sentry.models.avatars.user_avatar import UserAvatar
         from sentry.models.identity import Identity
-        from sentry.models.options.user_option import UserOption
         from sentry.models.organizationmembermapping import OrganizationMemberMapping
         from sentry.models.useremail import UserEmail
         from sentry.users.models.authenticator import Authenticator
+        from sentry.users.models.user_option import UserOption
 
         from_user_id = from_user.id
         to_user_id = to_user.id
