+++ src/sentry/users/services/user_option/impl.py
@@ -8,7 +8,7 @@ from django.db.models import QuerySet
 from sentry.api.serializers.base import Serializer
 from sentry.auth.services.auth import AuthenticationContext
 from sentry.hybridcloud.rpc.filter_query import FilterQueryDatabaseImpl, OpaqueSerializedResponse
-from sentry.models.options.user_option import UserOption
+from sentry.users.models.user_option import UserOption
 from sentry.users.services.user import RpcUser
 from sentry.users.services.user_option import RpcUserOption, UserOptionFilterArgs, UserOptionService
