+++ tests/sentry/api/endpoints/test_user_details.py
@@ -2,7 +2,6 @@ from django.test import override_settings
 from pytest import fixture
 
 from sentry.models.deletedorganization import DeletedOrganization
-from sentry.models.options.user_option import UserOption
 from sentry.models.organization import Organization, OrganizationStatus
 from sentry.models.organizationmember import OrganizationMember
 from sentry.models.userpermission import UserPermission
@@ -14,6 +13,7 @@ from sentry.testutils.hybrid_cloud import HybridCloudTestMixin
 from sentry.testutils.outbox import outbox_runner
 from sentry.testutils.silo import assume_test_silo_mode, control_silo_test
 from sentry.users.models.user import User
+from sentry.users.models.user_option import UserOption
 from sentry.users.models.userrole import UserRole
