+++ tests/sentry/api/endpoints/test_user_emails.py
@@ -1,10 +1,10 @@
 from django.urls import reverse
 
-from sentry.models.options.user_option import UserOption
 from sentry.models.useremail import UserEmail
 from sentry.testutils.cases import APITestCase
 from sentry.testutils.silo import control_silo_test
 from sentry.users.models.user import User
+from sentry.users.models.user_option import UserOption
 
 
 @control_silo_test
