+++ tests/sentry/api/endpoints/test_user_notification_email.py
@@ -1,7 +1,7 @@
-from sentry.models.options.user_option import UserOption
 from sentry.models.useremail import UserEmail
 from sentry.testutils.cases import APITestCase
 from sentry.testutils.silo import control_silo_test
+from sentry.users.models.user_option import UserOption
 
 
 class UserNotificationEmailTestBase(APITestCase):
