+++ tests/sentry/mail/activity/test_note.py
@@ -1,7 +1,6 @@
 from sentry.integrations.types import ExternalProviders
 from sentry.models.activity import Activity
 from sentry.models.notificationsettingoption import NotificationSettingOption
-from sentry.models.options.user_option import UserOption
 from sentry.notifications.notifications.activity.note import NoteActivityNotification
 from sentry.notifications.types import GroupSubscriptionReason
 from sentry.silo.base import SiloMode
@@ -9,6 +8,7 @@ from sentry.testutils.cases import ActivityTestCase
 from sentry.testutils.silo import assume_test_silo_mode
 from sentry.types.activity import ActivityType
 from sentry.types.actor import Actor
+from sentry.users.models.user_option import UserOption
 
 
 class NoteTestCase(ActivityTestCase):
